LayoutTests/http/tests/security/dataURL/xss-DENIED-from-data-url-sub-frame-to-data-url-sub-frame-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL data:text/html,<html><head><script>function fireSentinel() {if (window.layoutTestController)layoutTestController.globalFlag = true;}</script></head><body onload="fireSentinel();"><p>Inner iframe.</p><p id='accessMe'>Pass: Cross frame access from a sibling data: URL was denied.</p></body></html> from frame with URL data:text/html,<html><head><script>function loaded() {if (window.layoutTestController) {setTimeout(waitForFlag, 1);function waitForFlag() {if (!layoutTestController.globalFlag) {setTimeout(waitForFlag, 1);return;}performTest();}} else {document.getElementById('console').innerHTML = 'Click the button when all the subframes have finished loading.';}}function performTest() {try {parent.frames['firstSubFrame'].document.getElementById('accessMe').innerHTML = 'FAIL: Cross frame access from a sibling data: URL was allowed.';} catch (e) {}if (window.layoutTestController)layoutTestController.notifyDone();}</script></head><body onload="loaded();"><p>Inner iframe.</p><button onclick='performTest();'>Test</button><pre id='console'></pre></body></html>. Domains, protocols and ports must match.

This tests that a data: URL subframe can't access a sibling data: URL subframe.

  

--------
Frame: 'firstSubFrame'
--------
Inner iframe.

Pass: Cross frame access from a sibling data: URL was denied.

--------
Frame: 'secondSubFrame'
--------
Inner iframe.

Test
